Improving readability of code and split method in more classes to maintability
