+++ LevelEditor/data/levels/level_04.json
@@ -1,1 +1,1 @@
-{"id":"level_04","name":"level_04","background":"airadventurelevel3.png","ammo":"323","obstacles":2,"targets":1,"one_star":"42","two_stars":"054","three_stars":"66","cannon":{"id":"cannon","pos":{"x":65,"y":500}},"collidableLists":{"obstacleList":[{"id":"obstacle-0","pos":{"x":648,"y":201},"entity":{"type":"obstacle","name":"crate","height":"5","width":"5","texture":"crate-one.png","shape":"square","friction":"1","mass":"90","restitution":"0"}},{"id":"obstacle-2","pos":{"x":843,"y":506},"entity":{"type":"obstacle","name":"crate","height":"5","width":"5","texture":"crate-one.png","shape":"square","friction":"1","mass":"90","restitution":"0"}}],"targetList":[{"id":"target-1","pos":{"x":458,"y":154},"entity":{"type":"target","name":"bird","height":"1","width":"1","texture":"bird.png","shape":"square","friction":"1","mass":"50","restitution":"0","value":"100"}}]}}
+{"id":"level_04","name":"level_04","background":"airadventurelevel3.png","ammo":"323","one_star":"42","two_stars":"54","three_stars":"66","cannon":{"id":"cannon","pos":{"x":65,"y":500}},"collidableLists":{"obstacleList":[{"id":"obstacle-0","pos":{"x":648,"y":201},"entity":{"type":"obstacle","name":"crate","height":"5","width":"5","texture":"crate-one.png","shape":"square","friction":"1","mass":"90","restitution":"0"}},{"id":"obstacle-1","pos":{"x":843,"y":506},"entity":{"type":"obstacle","name":"crate","height":"5","width":"5","texture":"crate-one.png","shape":"square","friction":"1","mass":"90","restitution":"0"}}],"targetList":[{"id":"target-2","pos":{"x":458,"y":154},"entity":{"type":"target","name":"bird","height":"1","width":"1","texture":"bird.png","shape":"square","friction":"1","mass":"50","restitution":"0","value":"100"}}]}}
